perf(ui): Move annotated text useProjects inside subcomponent
We only use projects/organization when displaying [Filtered], the rest are static. By moving it lower we can reduce the number of useProjects on the page (assuming not all values are filtered).
