+++ static/app/components/events/meta/annotatedText/index.tsx
@@ -1,7 +1,3 @@
-import {useLocation} from 'sentry/utils/useLocation';
-import useOrganization from 'sentry/utils/useOrganization';
-import useProjects from 'sentry/utils/useProjects';
-
 import {AnnotatedTextErrors} from './annotatedTextErrors';
 import {AnnotatedTextValue} from './annotatedTextValue';
 
@@ -12,20 +8,9 @@ type Props = {
 };
 
 export function AnnotatedText({value, meta, className, ...props}: Props) {
-  const organization = useOrganization();
-  const location = useLocation();
-  const projectId = location.query.project;
-  const {projects} = useProjects();
-  const currentProject = projects.find(project => project.id === projectId);
-
   return (
     <span className={className} {...props}>
-      <AnnotatedTextValue
-        value={value}
-        meta={meta}
-        project={currentProject}
-        organization={organization}
-      />
+      <AnnotatedTextValue value={value} meta={meta} />
       <AnnotatedTextErrors errors={meta?.err} />
     </span>
   );
